Store all dates as UTC for SQLite
Dates are stored in SQLite as Unix timestamps. Dates are
represented in JavaScript as Unix timestamps. Thus, there need
not be any sort of conversions. Client-side timezones are
irrelevant; they should only affect the client-side display of
dates. The existing code is incorrect because it's storing
timezone-offset dates _as UTC dates_ (with the "Z" suffix).

This change simplifies the type conversion; a Date is stored
as an ISO-8601 date string, which is understood by both
JavaScript's `Date` constructor and SQLite.